// src/interfaces/IService.ts
export default interface IService {
    _id?: string
    code?: string
    serverName?: string
    price?: number
    cpu?: string
    ram?: string
    ssd?: string
    bandwidth?: string
    tranfer?: string
    ipv4?: string
    serverDefault?: boolean
    createdTime?: string
    discount?: number
    isCheck?: boolean
    expiryDateType: number
}